third_party/WebKit/LayoutTests/http/tests/inspector/console-cross-origin-iframe-logging-expected.txt

CONSOLE ERROR: Unable to post message to http://127.0.0.1:8000. Recipient has origin http://localhost:8000.

Tests that cross origin errors are logged with source url and line number.


Unable to post message to http://127.0.0.1:8000. Recipient has origin http://localhost:8000.
 console-cross-origin-iframe-logging.html:19
